requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-restful==0.3.10
python-dotenv==1.0.0
psycopg2-binary==2.9.9
click==8.1.7
pytest==7.4.0
sqlalchemy==2.0.23